requirements.txt
django>=1.10

# configuration
django-dotenv

# database
psycopg2

django-mptt
django-modeltranslation

# search
django-haystack
celery-haystack
django-haystack-es
elasticsearch
certifi

# cache
django-redis

# tasks
Celery

# data
django-import-export

# admin
django-modeladmin-reorder
django-postgres-utils

# media
Pillow
django-imagekit

# api
djangorestframework
markdown
django-filter
django-cors-headers
djangorestframework-csv
drf-haystack

# development
django-debug-toolbar

